Implemented list view
The menu is now shown as a list (tabs to be implemented): items are not
yet "styled" so it is ugly: but it works. The AdView has been moved to
the fragment so it behaves nicely with a collapsing toolbar.

The toolbar now uses a custom font for the app name (thanks Google
Fonts): this is to give the app some branding.